requirements.txt
fastapi==0.112.0
fastapi-login==1.10.2
fastapi-session==0.2.7
google-generativeai==0.7.2
itsdangerous==2.2.0
jinja2==3.1.4
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.9
pypdf2==3.0.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
sqlalchemy==2.0.35
uvicorn==0.30.6
